add skeleton for afsk1200
add skeleton for afsk1200
