+++ directdemod/constants.py
@@ -3,12 +3,12 @@
 ## IQ.wav settings
 IQ_FREQOFFSET = 30000
 IQ_SDRSAMPRATE = 2.048e6
-IQ_FMBW = 60000
 
 ## Processing settings
 PROC_CHUNKSIZE = 20000000
 
 ## NOAA settings
+NOAA_FMBW = 60000
 NOAA_AUDSAMPRATE = 20800
 NOAA_FREQ = 137620000
 NOAA_CRUDESYNCSAMPRATE = 40960
